+++ client/src/components/header/Header.tsx
@@ -1,20 +1,10 @@
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import { ReactComponent as LogoSvg } from '@/assets/logos/logo.svg';
 import { ReactComponent as SearchIconSvg } from '@/assets/icons/search_icon.svg';
-import Button from '../ui/Button';
-import { useAppSelector } from '@/hooks/useReducer';
-import { profile } from '@/assets/mypage/index';
+import UserNav from './UserNav';
 
 function Header() {
-  const navigate = useNavigate();
-  const isLogin = useAppSelector((state) => state.user.isLogin);
-  const userData = useAppSelector((state) => state.user.data);
-
-  const onProjectAddClick = () => {
-    navigate('/project/add');
-  };
-
   return (
     <header className='flex-center h-62pxr'>
       <div className='flex items-center justify-between w-full max-w-7xl'>
@@ -29,30 +19,7 @@ function Header() {
           />
           <SearchIconSvg className='absolute right-20pxr top-8pxr' />
         </div>
-        <div className='flex-center ellipsis'>
-          {isLogin ? (
-            <Link
-              to='/mypage'
-              className='overflow-hidden bg-gray-400 rounded-full h-35pxr w-35pxr mr-20pxr'
-            >
-              <img src={userData?.userImg ? userData.userImg : profile} alt='유저' />
-            </Link>
-          ) : (
-            <>
-              <Link to='/users/login' className='text-gray-800 px-10pxr hover:text-purple-300'>
-                로그인
-              </Link>
-              <Link to='/users/signup' className='text-gray-800 px-10pxr hover:text-purple-300'>
-                회원가입
-              </Link>
-            </>
-          )}
-          <Button
-            onClick={onProjectAddClick}
-            text={'프로젝트 만들기'}
-            style={'text-sm ml-7pxr h-35pxr px-20pxr py-8pxr'}
-          />
-        </div>
+        <UserNav />
       </div>
     </header>
   );
