+++ client/src/components/kakaoshare/ShareModal.tsx
@@ -46,7 +46,8 @@ function ShareModal({ onModalClosed, modalData }: Props) {
       <div className='flex w-full gap-10pxr h-37pxr'>
         <div className='relative flex w-full h-full'>
           <input
-            defaultValue={window.location.href}
+            value={window.location.href}
+            readOnly
             className='border-[1.5px] rounded w-full border-gray-300 ellipsis pl-10pxr pr-55pxr'
           />
           <button
